feat(controllers): implement admin router, comment logic and security checks

+++ controllers/userRouter.js
@@ -1,6 +1,9 @@
 const userRouter = require('express').Router()
 const User = require('../models/user')
 const bcrypt = require ('bcrypt')
+const { userExtractor } = require('../utils/middleware')
+
+userRouter.use(userExtractor)
 
 userRouter.get ('/', async (_req, res, next) => {
 
@@ -61,7 +64,8 @@ userRouter.post('/', async (req, res, next) => {
     const user = new User({
       username,
       name,
-      passwordHash
+      passwordHash,
+      role: 'user'
     })
 
     const newUser = await user.save()
@@ -76,21 +80,76 @@ userRouter.post('/', async (req, res, next) => {
 
 
 userRouter.delete ('/:id', async (req, res, next) => {
-
+  const user = req.user
   const { id } = req.params
-  try
-  {
-    await User
-      .findByIdAndDelete (id)
 
+  try {
+
+    const deletedUser = await User.findById(id)
+
+    if (!deletedUser)
+      return res
+        .status(404)
+        .json({ error: 'user not found' })
+
+    const isAdmin = user.role === 'admin'
+    const isSelf = deletedUser._id.toString()=== user._id.toString()
+
+    if (!isAdmin && !isSelf)
+      return res.status(401).json({ error: 'unauthorized' })
+
+    await User
+      .findByIdAndDelete(id)
     return res
       .status(204)
       .end()
 
-  } catch (error){
-    next (error)
+  } catch (error) {
+    next(error)
   }
 
 })
+
+userRouter.put ('/:id', async (req, res, next) => {
+
+  const user  = req.user
+  const { id } = req.params
+  const { username, name } = req.body
+
+  try
+  {
+    const updatedUser = await User.findById(id)
+
+    if (!updatedUser)
+      return res
+        .status(404)
+        .json({ error: 'user not found' })
+
+    const itSelf = updatedUser.id === user.id
+    const isAdmin = user.role === 'admin'
+
+    if (!itSelf && !isAdmin)
+      return res
+        .status(401)
+        .json({ error: 'you dont have permission' })
+
+
+    const updatedUserData = {
+      username: username || updatedUser.username,
+      name: name || updatedUser.name
+    }
+
+    const updated = await User
+      .findByIdAndUpdate(id, updatedUserData, { new: true })
+
+    return res
+      .status(200)
+      .json(updated)
+  }
+  catch (error){
+    next (error)
+  }
+})
+
 
 module.exports=userRouter
